requirements.txt
flask
python-dotenv
pypdf2
langchain>=0.2.0
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
faiss-cpu
numpy<2.0.0
torch --index-url https://download.pytorch.org/whl/cpu
sentence-transformers
markdown
gunicorn